Dodanih več opomb in pojasnil
